Added Auth Checks to Image Routes

+++ frontend/src/app/api/user/images/deleteImage.ts
@@ -6,13 +6,11 @@ import { HTTP_RESPONSES } from "@/lib/constants/httpResponses";
 import { NextResponse } from "next/server";
 import { ObjectId } from "mongodb";
 import { isValidCollectionName } from "@/lib/utils/typeValidation/general";
+import { currentUser } from "@clerk/nextjs/server";
 
 export async function DELETE(req: Request) {
   try {
     const { imageId, collectionName } = await req.json()
-
-    console.log(imageId)
-    console.log(collectionName)
 
     if (!imageId || !isValidCollectionName(collectionName)) {
       return NextResponse.json(
@@ -21,7 +19,15 @@ export async function DELETE(req: Request) {
       );
     }
 
-    const { bucket } = await setupGridFS();
+    const { bucket, db } = await setupGridFS();
+
+    const clerkUser = await currentUser();
+
+    if (!clerkUser || !clerkUser.id) {
+      return NextResponse.json({ message: "User not found" }, { status: 404 });
+    }
+
+    const userProfile = await db.collection('users').findOne({ clerkId: clerkUser.id });
 
     const imageIdAsObjectId = ObjectId.createFromHexString(imageId);
     const file = await bucket.find({ _id: imageIdAsObjectId }).toArray();
@@ -33,11 +39,13 @@ export async function DELETE(req: Request) {
       );
     }
 
+    if (file[0].metadata?.creatorId.toString() !== userProfile?._id.toString()) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 403 });
+    }
+
     const sourceId = file[0].metadata?.sourceId;
 
     await bucket.delete(file[0]._id);
-
-    const db = await getDB();
 
     await db.collection(collectionName).updateOne({
       _id: ObjectId.createFromHexString(sourceId)
